% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellAssign.R
\name{cellAssign}
\alias{cellAssign}
\title{Annotate all cells with contour level of cell type-specific density.}
\usage{
cellAssign(
  spe,
  coi = NULL,
  name_to = paste0("at_level_", janitor::make_clean_names(coi)),
  NA_level = "0",
  levels = NULL,
  assign = c("contour", "roi")
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{name_to}{Colname in colData(spe) to store the annotation.}

\item{NA_level}{Label for cells not falling in any of the regions. Default to 0.}

\item{levels}{Factor levels.}
}
\value{
A SpatialExperiment object. An extra column is added to the colData.
}
\description{
Annotate all cells with contour level of cell type-specific density.
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

coi <- "Breast cancer"

spe <- getContour(spe, coi = coi)

spe <- contourAssign(spe, coi = coi)

}
