% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsInRegion.R
\name{cellsInRegion}
\alias{cellsInRegion}
\title{Check which cells are in which regions}
\usage{
cellsInRegion(spe, sf, name_to, NA_level = NULL, levels = NULL)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{sf}{List or an sf object that represents a region or an ROI.}

\item{name_to}{Colname in colData(spe) to store the annotation.}

\item{NA_level}{Label for cells not falling in any of the regions.}

\item{levels}{Factor levels.}
}
\value{
An spe object.
}
\description{
Check which cells are in which regions
}
