% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{compute_density}
\alias{compute_density}
\title{Perform kernel density estimation on SpatialExperiment}
\usage{
compute_density(
  spe,
  bandwidth = NULL,
  weights = NULL,
  kernel = c("gaussian", "epanechnikov", "quartic", "disc"),
  scale = 10000,
  ngrid_x = 100,
  ngrid_y = NULL,
  grid_length_x = NULL,
  grid_length_y = NULL,
  xlim,
  ylim
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{bandwidth}{The smoothing bandwidth. By default performing automatic bandwidth}

\item{weights}{Optional weights to be attached to the points.}

\item{kernel}{The smoothing kernel.Options are gaussian, epanechnikov, quartic or disc.}

\item{scale}{Scaling factor for density.}

\item{ngrid_x}{Number of grids in the x-direction.}

\item{ngrid_y}{Number of grids in the y-direction.}

\item{grid_length_x}{Grid length in the x-direction.}

\item{grid_length_y}{Grid length in the y-direction.}

\item{xlim}{The range of the x-coordinates of the image.}

\item{ylim}{The range of the y-coordinates of the image.

selection using cross-validation using function spatstat.explore::bw.diggle.}
}
\value{
A SpatialExperiment object.
}
\description{
Perform kernel density estimation on SpatialExperiment
}
