% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDensity.R
\name{computeDensity}
\alias{computeDensity}
\title{Perform kernel density estimation on SpatialExperiment}
\usage{
computeDensity(
  spe,
  mode = c("pixels", "points"),
  kernel = c("gaussian", "epanechnikov", "quartic", "disc"),
  bandwidth = NULL,
  weights = NULL,
  ngrid.x = NULL,
  ngrid.y = NULL,
  grid.length.x = 100,
  grid.length.y = NULL,
  xlim = NULL,
  ylim = NULL,
  diggle = FALSE
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{mode}{Choose either points or pixels. Specifying
whether to compute the density at a grid pixel location or at at point.}

\item{kernel}{The smoothing kernel. Options are gaussian, epanechnikov,
quartic or disc.}

\item{bandwidth}{The smoothing bandwidth. By default performing automatic
bandwidth selection using cross-validation using function
spatstat.explore::bw.diggle.}

\item{weights}{Optional weights to be attached to the points.}

\item{ngrid.x}{Number of grids in the x-direction.}

\item{ngrid.y}{Number of grids in the y-direction.}

\item{grid.length.x}{Grid length in the x-direction.
Default to 100 (micron).}

\item{grid.length.y}{Grid length in the y-direction.}

\item{xlim}{The range of the x-coordinates of the image.}

\item{ylim}{The range of the y-coordinates of the image.}

\item{diggle}{Logical. If TRUE, use the Jones-Diggle improved edge
correction. See spatstat.explore::density.ppp() for details.}
}
\value{
Output from spatstat.explore::density.ppp.
}
\description{
Perform kernel density estimation on SpatialExperiment
}
