% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour2sf.R
\name{contour2sf}
\alias{contour2sf}
\title{Draw a contour region on some density level}
\usage{
contour2sf(spe, contour, cutoff)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{contour}{Name in metadata.}

\item{cutoff}{A numeric scalar specifying the density cutoff.}
}
\value{
An sf object of the contour region of the specified level.
}
\description{
Draw a contour region on some density level
}
