% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNbrsGrid.R
\name{coord_hash}
\alias{coord_hash}
\title{Hash two 15-bytes signed integers into one 32-bytes integer.}
\usage{
coord_hash(a, b)
}
\arguments{
\item{a, b}{Integer vectors of same lengths. Can be negative.}
}
\description{
Hash two 15-bytes signed integers into one 32-bytes integer.
}
\details{
Should work for a,b in range (-2^14, 2^14-1) which is good enough for our
purpose.

Integer in R is 32-bytes but reserve 1 byte for NA
}
