% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_roi_by_dens.R
\name{find_roi_by_dens}
\alias{find_roi_by_dens}
\title{Find ROIs based on cell type-specific densities via graph-based method.}
\usage{
find_roi_by_dens(
  spe,
  coi = NULL,
  probs = 0.85,
  method = c("walktrap", "connected"),
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{probs}{A numeric vector. The threshold of proportion that used to
filter grid by density.}

\item{method}{The community dectection method to be used, either walktrap or connected.}

\item{...}{Other parameters that passed to walktrap.community.}
}
\value{
A SpatialExperiment object.
}
\description{
Find ROIs based on cell type-specific densities via graph-based method.
}
\examples{

data("xenium_bc_spe")

spe_dens <- grid_dens_by_coi(spe, coi = c("Breast cancer", "Fibroblasts"))

spe_rois <- find_roi_by_dens(spe_dens, coi = c("Breast cancer", "Fibroblasts"), 
method = "walktrap")

}
