% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContourRegions.R
\name{getContourRegions}
\alias{getContourRegions}
\title{Calculate areas between every two density levels}
\usage{
getContourRegions(spe, contour_name)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{contour_name}{Name of contour in spe@metadata}
}
\value{
A list of sf objects, each representing the region between
two contour density levels.
}
\description{
Calculate areas between every two density levels
}
