% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_dens_by_coi.R
\name{grid_dens_by_coi}
\alias{grid_dens_by_coi}
\title{Perform kernel density estimation on SpatialExperiment for cell types of interest}
\usage{
grid_dens_by_coi(
  spe,
  coi,
  coi.cn = "cell_type",
  ngrid_x = 100,
  ngrid_y = 100,
  bandwidth = 50
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{coi.cn}{A character vector. Column name of the cell types.
By default is cell_type.}

\item{ngrid_x}{Number of grids in the x-direction.}

\item{ngrid_y}{Number of grids in the y-direction.}

\item{bandwidth}{Smoothing parameter}
}
\value{
A SpatialExperiment object.
}
\description{
Perform kernel density estimation on SpatialExperiment for cell types of interest
}
\examples{

data("xenium_bc_spe")

spe_dens <- grid_dens_by_coi(spe, coi = c("Breast cancer", "Fibroblasts"))


}
