% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbrs2igraph.R
\name{nbrs2igraph}
\alias{nbrs2igraph}
\title{Convert nbrs (in spe@metadata$nbrs) into igraph's graph}
\usage{
nbrs2igraph(nbrs, directed = FALSE)
}
\description{
Convert nbrs (in spe@metadata$nbrs) into igraph's graph
}
\details{
nbrs should be a list containing index & weight
}
