% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_density.R
\name{plot_density}
\alias{plot_density}
\title{Plot grid-based density.}
\usage{
plot_density(spe, q.tile = 0.8)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{q.tile}{Numeric value between 0 and 1, used for filtering uninformative grid, default is 0.8.}
}
\value{
A ggplot object.
}
\description{
Plot grid-based density.
}
\examples{

data("xenium_bc_spe")

spe <- compute_density(spe)

plot_density(spe)
}
