% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_roi.R
\name{plot_roi}
\alias{plot_roi}
\title{Plot ROIs on spatial.}
\usage{
plot_roi(
  spe,
  coi = NULL,
  col.cn = "cell_type",
  n = 30,
  threshold = 0,
  showlegend = FALSE,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{col.cn}{A character vector. Column name of the cell types.
By default is cell_type.}

\item{n}{Integer value. The number of distinct color to be generated, default is 30.}

\item{threshold}{Integer value. The threshold (number of grid) used to find large ROIs.}

\item{showlegend}{Logical. Show legend or not.}

\item{...}{Aesthetic mappings to pass to \code{ggplot2::aes_string()} for point.}
}
\value{
A ggplot object.
}
\description{
Plot ROIs on spatial.
}
\examples{

data("xenium_bc_spe")

spe_dens <- grid_dens_by_coi(spe, coi = c("Breast cancer", "Fibroblasts"))

spe_rois <- find_roi_by_dens(spe_dens, coi = c("Breast cancer", "Fibroblasts"), 
method = "walktrap", steps = 5)

plot_roi(spe_rois, coi = c("Breast cancer", "Fibroblasts"), size = 0.01, alpha = 0.2)

}
