% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_spatial.R
\name{plot_spatial}
\alias{plot_spatial}
\title{Plot cells based on spatial coordinates.}
\usage{
plot_spatial(spe, reverseY = FALSE, n = 30, ...)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{reverseY}{Reverse y coordinates.}

\item{n}{Integer value. The number of distinct color to be generated, default is 30.}

\item{...}{Aesthetic mappings to pass to \code{ggplot2::aes_string()}.}
}
\value{
A ggplot object.
}
\description{
Plot cells based on spatial coordinates.
}
\examples{

data("xenium_bc_spe")

plot_spatial(spe, shape = ".", color = cell_type)

}
