% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensCor.R
\name{plotDensCor}
\alias{plotDensCor}
\title{Plot density correlation between two cell types}
\usage{
plotDensCor(
  spe,
  celltype1 = NULL,
  celltype2 = NULL,
  by.roi = TRUE,
  probs = 0.85,
  fit = c("spline", "linear"),
  df = 3,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{celltype1}{Cell type 1 to compare.}

\item{celltype2}{Cell type 2 to compare.}

\item{by.roi}{Logical. Plot facet by ROIs or not. Default to TRUE.}

\item{probs}{A numeric scalar. The threshold of proportion that used to
filter grids by density when ROIs have not been identified previously.
Ignored if 'roi' is present in the 'metadata' component of spe. Default to 0.85.}

\item{fit}{Character. Options are "spline" and "linear".}

\item{df}{Integer. Degrees of freedom of the spline fit.
Default to 3 (i.e., a cubic spline fit).}

\item{...}{aesthetic mappings to pass to \code{ggplot2::aes()}.}
}
\value{
A ggplot object.
}
\description{
Plot density correlation between two cell types
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

plotDensCor(spe, celltype1 = "Breast cancer", celltype2 = "Fibroblasts")

}
