% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLISAscatter.R
\name{plotLISAscatter}
\alias{plotLISAscatter}
\title{Scatterplot for local moran's I}
\usage{
plotLISAscatter(
  lisa,
  quadrant.count = TRUE,
  text.size = 11/.pt,
  xlab = "Data",
  ylab = "Spatial Lag"
)
}
\arguments{
\item{lisa}{A list obtained from \link[scider]{localMoran}}

\item{quadrant.count}{Whether to count values at each quadrant (Low-Low,
Low-High, High-High, High-Low)}

\item{text.size}{Numeric for text size of quadrant.count}

\item{xlab}{label for the x-axis}

\item{ylab}{label for the y-axis}
}
\value{
A ggplot object.
}
\description{
plot result obtained from localMoran()
}
\examples{

data("xenium_bc_spe")
dat <- spe$total_counts
spe <- findNbrsSpatial(spe,k=50)
res <- localMoran(spe,data1=dat,at="cell")
plotLISAscatter(res)
}
