% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModStat.R
\name{plotModStat}
\alias{plotModStat}
\title{Plot model statistics using heatmap.}
\usage{
plotModStat(spe, stats = c("coefficient", "t"), roi = NULL)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{stats}{Character value. Choose either coefficient or t.
Coefficient by default.}

\item{roi}{Specify the ROI to be plotted. Leave it as default if modelling
is not done by per ROI or just want to first ROI to be plotted.}
}
\value{
A pheatmap object.
}
\description{
Plot model statistics using heatmap.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts", "B cells", "T cells")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

spe <- corrDensity(spe, cal_all = TRUE)

plotModStat(spe)

}
