% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readST.R
\name{readProseg}
\alias{readProseg}
\title{Read Proseg V2 output into spe}
\usage{
readProseg(
  dir,
  sample_id = "sample01",
  count = "expected-counts\\\\.(csv|mtx)\\\\.gz",
  coord = "cell-metadata.csv.gz",
  gene = "gene-metadata.csv.gz",
  coordNames = c("centroid_x", "centroid_y", "centroid_z")
)
}
\arguments{
\item{dir}{directory containing the Proseg files}

\item{sample_id}{Name of the sample.}

\item{count}{Name of the file with the count assay.}

\item{coord}{Name of the file with the tissue coordinates}

\item{gene}{Name of the file with the gene metadata. This}

\item{coordNames}{Name of the coordinates for the spe}
}
\description{
Read Proseg V2 output into spe
}
\details{
This does not work on zarr file output of proseg V3
}
