% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realignVisium.R
\name{realignVisium}
\alias{realignVisium}
\title{Scale and straighten out Visium coordinates}
\usage{
realignVisium(spe, distPoint = 100)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{distPoint}{Numeric. Desired point to point distance.}
}
\description{
Scale and straighten out Visium coordinates
}
\details{
This function rescale the distance between points to 100um (or other
value) to matchthe real distance. In addition, Visium spots have a slight
tilt to them which this function will also fix
}
