% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/realignVisium.R
\name{realignVisiumHD}
\alias{realignVisiumHD}
\title{Scale and straighten out VisiumHD coordinates}
\usage{
realignVisiumHD(spe, distPoint = NULL)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{distPoint}{Numeric. Desired point to point distance. If NULL, will try
to determine the bin level of spe and use that.}
}
\description{
Scale and straighten out VisiumHD coordinates
}
\details{
This function rescale the distance between points to 8um (or other
value) to matchthe real distance. In addition, Visium spots have a slight
tilt to them which this function will also fix
}
