% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_region.R
\name{select_region}
\alias{select_region}
\title{Select region of interest from plot}
\usage{
select_region(spe)
}
\arguments{
\item{spe}{A SpatialExperiment object.}
}
\value{
A matrix.
}
\description{
Select region of interest from plot
}
\examples{

data("xenium_bc_spe")

spe_b <- spe[,colData(spe)$cell_type == "B cells"]

select_region(spe_b)

}
