% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{[,SpatialExperiment,ANY,ANY,ANY-method}
\alias{[,SpatialExperiment,ANY,ANY,ANY-method}
\title{Subset for grid level analysis}
\usage{
\S4method{[}{SpatialExperiment,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A SpatialExperiment object.}

\item{i}{row indices for subsetting.}

\item{j}{col indices for subsetting.}

\item{...}{further arguments to be passed to or from other methods.}

\item{drop}{passed on to [ indexing operator.}
}
\value{
A SpatialExperiment object.
}
\description{
Overwrite the default SpatialExperiment subsetting method to ensure
'grid_density' is also subsetted if 'gridLevelAnalysis' is TRUE (1 polygon 1 spot)
}
