% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{update_bound}
\alias{update_bound}
\title{Update the x,y limits of a plot}
\usage{
update_bound(p, x = NULL, y = NULL)
}
\arguments{
\item{p}{A ggplot() object}

\item{x, y}{Vectors of new x and y limits.}
}
\value{
a ggplot object
}
\description{
Update the x,y limits of a plot
}
