% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scider-package.R
\docType{data}
\name{xenium_bc_spe}
\alias{xenium_bc_spe}
\alias{spe}
\title{Description of the scider example datasets}
\format{
A SpatialExperiment object
}
\usage{
data("xenium_bc_spe")
}
\value{
A SpatialExperiment object
}
\description{
scider-package has 1 datasets: \itemize{
\item xenium_bc_spe Example test spatial transcriptomics data
in SpatialExperiment format. This test data is randomly subsetting
from a publicly available 10X Xenium breast cancer data.
Source data: https://www.10xgenomics.com/resources/datasets/xenium-ffpe-human-breast-with-custom-add-on-panel-1-standard
}
}
\examples{
data(xenium_bc_spe)
}
\keyword{internal}
