% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allocateCells.R
\name{allocateCells}
\alias{allocateCells}
\title{Annotate all cells with contour level of cell type-specific density.}
\usage{
allocateCells(spe, to.roi = TRUE, to.contour = TRUE)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{to.roi}{Logical. Whether to allocate cells to ROIs.}

\item{to.contour}{Logical. Whether to allocate cells to contour levels.}
}
\value{
A SpatialExperiment object. An extra column is added to the colData.
}
\description{
Annotate all cells with contour level of cell type-specific density.
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

coi <- "Breast cancer"

spe <- findROI(spe, coi = coi)

spe <- allocateCells(spe)

}
