% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellsInRegion.R
\name{cellsInRegion}
\alias{cellsInRegion}
\title{Check which cells are in which regions}
\usage{
cellsInRegion(spe, region, name_to, NA_level = "0", levels = NULL)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{region}{List or an sf object that represents a region or an ROI.}

\item{name_to}{Colname in colData(spe) to store the annotation.}

\item{NA_level}{Label for cells not falling in any of the regions.
Default to 0.}

\item{levels}{Factor levels.}
}
\value{
A SpatialExperiment object. The region information of each cell is
stored in the colData.
}
\description{
Check which cells are in which regions
}
