% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_density.R
\name{compute_density}
\alias{compute_density}
\title{Perform kernel density estimation on SpatialExperiment}
\usage{
compute_density(
  spe,
  mode = c("points", "pixels"),
  bandwidth = NULL,
  weights = NULL,
  kernel = c("gaussian", "epanechnikov", "quartic", "disc"),
  grid_size = 100,
  xlim = NULL,
  ylim = NULL,
  den_scaler = 10000
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{mode}{Choose either points or pixels. Specifying whether to compute the
density at a grid pixel location or at at point.}

\item{bandwidth}{The smoothing bandwidth. By default performing automatic bandwidth
selection using cross-validation using function spatstat.explore::bw.diggle.}

\item{weights}{Optional weights to be attached to the points.}

\item{kernel}{The smoothing kernel.Options are gaussian, epanechnikov, quartic or disc.}

\item{grid_size}{Grid size to use when the mode is set to pixels.}

\item{den_scaler}{Numeric value, used for scaling density value, default is 1e4.}
}
\value{
A SpatialExperiment object.
}
\description{
Perform kernel density estimation on SpatialExperiment
}
\examples{

data("xenium_bc_spe")

compute_density(spe)

}
