% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDensity.R
\name{computeDensity}
\alias{computeDensity}
\title{Perform kernel density estimation on SpatialExperiment}
\usage{
computeDensity(
  spe,
  mode = "pixels",
  kernel = "gaussian",
  bandwidth = NULL,
  weights = NULL,
  scale = 10000,
  ngrid_x = 100,
  ngrid_y = NULL,
  grid_length_x = NULL,
  grid_length_y = NULL,
  xlim = NULL,
  ylim = NULL
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{mode}{Choose either points or pixels. Specifying whether to compute the
density at a grid pixel location or at at point.}

\item{kernel}{The smoothing kernel. Options are gaussian, epanechnikov, quartic or disc.}

\item{bandwidth}{The smoothing bandwidth. By default performing automatic bandwidth
selection using cross-validation using function spatstat.explore::bw.diggle.}

\item{weights}{Optional weights to be attached to the points.}

\item{scale}{A munmeric vector to scale the density values.}

\item{ngrid_x}{Number of grids in the x-direction. Default to 100.}

\item{ngrid_y}{Number of grids in the y-direction.}

\item{grid_length_x}{Grid length in the x-direction.}

\item{grid_length_y}{Grid length in the y-direction.}

\item{xlim}{The range of the x-coordinates of the image.}

\item{ylim}{The range of the y-coordinates of the image.}
}
\value{
Output from spatstat.explore::density.ppp.
}
\description{
Perform kernel density estimation on SpatialExperiment
}
\examples{

data("xenium_bc_spe")

dens <- computeDensity(spe)

}
