% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDensityHex.R
\name{computeDensityHex}
\alias{computeDensityHex}
\title{Perform kernel density estimation on SpatialExperiment}
\usage{
computeDensityHex(
  spe,
  kernel = c("gaussian"),
  bandwidth = NULL,
  weights = NULL,
  ngrid.x = NULL,
  grid.length.x = 100,
  xlim = NULL,
  ylim = NULL,
  diggle = FALSE,
  isVisium = F
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{kernel}{The smoothing kernel. Options are gaussian, epanechnikov,
quartic or disc. ONLY GAUSSIAN IS IMPLEMENTED}

\item{bandwidth}{The smoothing bandwidth. By default performing automatic
bandwidth selection using cross-validation using function
spatstat.explore::bw.diggle.}

\item{weights}{Optional weights to be attached to the points.}

\item{ngrid.x}{Number of grids in the x-direction.}

\item{grid.length.x}{Grid length in the x-direction.
Default to 100 (micron).}

\item{xlim}{The range of the x-coordinates of the image.}

\item{ylim}{The range of the y-coordinates of the image.}

\item{diggle}{Logical. If TRUE, use the Jones-Diggle improved edge}

\item{isVisium}{Logical. If TRUE, fit hexagonal grids to Visium spots.
correction. See spatstat.explore::density.ppp() for details.}
}
\value{
Output from spatstat.explore::density.ppp.
}
\description{
Perform kernel density estimation on SpatialExperiment
}
