% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corDensity.R
\name{corDensity}
\alias{corDensity}
\title{Test for density correlation between two cell types.}
\usage{
corDensity(spe, coi = NULL, trace = TRUE)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{Character vector for cell types of interest for density
correlation analysis. Default is NULL, which is to consider all cell types
previously calculated in the gridDensity() step.}

\item{trace}{Logical. If TRUE (default), print process pf testing.}
}
\value{
A DataFrame containing the testing results.
}
\description{
Test for density correlation between two cell types.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts", "B cells", "T cells")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

result <- corDensity(spe)

}
