% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrDensity.R
\name{corrDensity}
\alias{corrDensity}
\title{Test for density correlation between two cell types.}
\usage{
corrDensity(spe, by.roi = TRUE)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{by.roi}{Logical. If TRUE (default), then return the testing results at ROI level.
If FALSE, then combine the testing results across all ROIs.}
}
\value{
A DataFrame containing the testing results.
}
\description{
Test for density correlation between two cell types.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts", "B cells", "T cells")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

result <- corrDensity(spe)

}
