% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNbrsSNN.R
\name{findNbrsSNN}
\alias{findNbrsSNN}
\title{Construct a SNN neighbour list from assay.}
\usage{
findNbrsSNN(
  spe,
  assay = NULL,
  dimred = "PCA",
  n_dimred = NULL,
  k = 20,
  BNPARAM = BiocNeighbors::AnnoyParam(),
  type = c("jaccard", "rank", "number"),
  prune = NULL,
  nbrs_name = NULL,
  cpu_threads = 6
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{assay}{Name of assay for clustering. Incompatible with dimred.}

\item{dimred}{Name of the dimensionality reduction (e.g. PCA) for clustering.
Incompatible with assay}

\item{n_dimred}{Integer scalar or vector specifying the dimensions to use if
dimred is specified. Defaults to NULL, which uses all available dimensions of
the reduced dim (e.g. all PCs produced by runPCA).}

\item{k}{Integer scalar for number of nearest neighbors to find. Default to 20.}

\item{BNPARAM}{\link[BiocNeighbors]{BiocNeighborParam} object specifying the nearest neighbor algorithm. Default is Annoy.}

\item{type}{Type of weighting scheme for shared neighbors. Options are jaccard
(default), rank, and number. type="rank" is defined in Xu and Su (2015).}

\item{prune}{Edges with SNN weight at or below this value are removed before
clustering. Defaults to NULL, which applies a jaccard cutoff of 0.1 for "jaccard",
converted to the equivalent shared-neighbour count (2(k+1)*0.1/1.1) for "number".
type="rank" is not pruned by default. Set to 0 to disable.}

\item{nbrs_name}{Name of the neighbour list to be stored in spe. Default to be
assay/dimred + "_snn".}

\item{cpu_threads}{Number of cpu threads for parallel computation.}
}
\value{
A spe with the clusters stored in \link[SingleCellExperiment]{reducedDims}.

A SpatialExperiment object
}
\description{
Construct a SNN neighbour list from assay.
}
\details{
Construct a SNN neighbour list using either the spe's assay or reduced
dimension and store it in spe@metadata$nbrs$cell

neighbour list contain
}
\examples{

data("xenium_bc_spe")
spe <- runPCA(spe)
spe <- findNbrsSNN(spe,dimred="PCA")
}
