% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findROI.R
\name{findROI}
\alias{findROI}
\title{Find ROIs based on cell type-specific densities via graph-based method.}
\usage{
findROI(
  spe,
  coi,
  probs = 0.85,
  ngrid.min = 20,
  method = "walktrap",
  diag.nodes = FALSE,
  sequential.roi.name = TRUE,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{probs}{A numeric scalar. The threshold of proportion that used to
filter grid by density. Default to 0.85.}

\item{ngrid.min}{An integer. The minimum number of grids required for
defining a ROI. Default to 20.}

\item{method}{The community dectection method to be used, either walktrap or
connected. Default to walktrap, can be abbreviated.}

\item{diag.nodes}{Logical. Set this to TRUE to allow diagonal grid points
to be adjacent nodes.}

\item{sequential.roi.name}{Logical. Set this to FALSE if you want the
original ROI name before
filtering are retained.}

\item{...}{Other parameters that passed to walktrap.community when method =
"walktrap".}
}
\value{
A SpatialExperiment object.
}
\description{
Find ROIs based on cell type-specific densities via graph-based method.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

}
