% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContour.R
\name{getContour}
\alias{getContour}
\title{Get contour from density}
\usage{
getContour(spe, coi, bins = NULL, binwidth = NULL, breaks = NULL)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{bins}{}

\item{binwidth}{}

\item{breaks}{}
}
\value{
An sf object of the contour region of the specified level.
}
\description{
Get contour from density
}
