% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getContourRegions.R
\name{getContourRegions}
\alias{getContourRegions}
\title{Calculate areas between every two density levels}
\usage{
getContourRegions(spe, coi)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}
}
\value{
A list of sf objects, each representing the region between two contour density levels.
}
\description{
Calculate areas between every two density levels
}
\examples{
data("xenium_bc_spe")

spe <- gridDensity(spe)

coi <- "Breast cancer"

spe <- getContour(spe, coi = coi)

region <- getContourRegions(spe, coi = coi)

}
