% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2sf.R
\name{grid2sf}
\alias{grid2sf}
\title{Combine grids in each ROI to a sf region}
\usage{
grid2sf(spe)
}
\arguments{
\item{spe}{A SpatialExperiment object.}
}
\value{
List of ROIs saved as sf objects.
}
\description{
Combine grids in each ROI to a sf region
}
