% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid2sf.R
\name{grid2sf}
\alias{grid2sf}
\title{Combine grids in each ROI to a sf region}
\usage{
grid2sf(spe, ngrid = 20)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{ngrid}{Integer. The threshold (minimum number of grids) used to filter small ROIs. Default to 20.}
}
\value{
List of ROIs saved as sf objects.
}
\description{
Combine grids in each ROI to a sf region
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

coi <- "Breast cancer"

spe <- findROI(spe, coi = coi, probs = 0.85)

rois <- grid2sf(spe, ngrid = 20)

}
