% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridDensity.R
\name{gridDensity}
\alias{gridDensity}
\title{Perform kernel density estimation on SpatialExperiment for
cell types of interest}
\usage{
gridDensity(
  spe,
  id = "cell_type",
  coi = NULL,
  kernel = "gaussian",
  bandwidth = NULL,
  ngrid.x = 100,
  ngrid.y = NULL,
  grid.length.x = NULL,
  grid.length.y = NULL,
  diggle = FALSE
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{id}{A character. The name of the column of colData(spe) containing
the cell type identifiers. Set to cell_type by default.}

\item{coi}{A character vector of cell types of interest (COIs).
Default to all cell types.}

\item{kernel}{The smoothing kernel. Options are "gaussian",
"epanechnikov", "quartic" or "disc".}

\item{bandwidth}{The smoothing bandwidth. By default performing
automatic bandwidth selection using cross-validation using
function spatstat.explore::bw.diggle.}

\item{ngrid.x}{Number of grids in the x-direction. Default to 100.}

\item{ngrid.y}{Number of grids in the y-direction.}

\item{grid.length.x}{Grid length in the x-direction.}

\item{grid.length.y}{Grid length in the y-direction.}

\item{diggle}{Logical. If TRUE, use the Jones-Diggle improved edge
correction. See spatstat.explore::density.ppp() for details.}
}
\value{
A SpatialExperiment object. Grid density estimates for
all cell type of interest are stored in spe@metadata$grid_density.
Grid information is stored in spe@metadata$grid_info
}
\description{
Perform kernel density estimation on SpatialExperiment for
cell types of interest
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

}
