% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridDensity.R
\name{gridDensity}
\alias{gridDensity}
\title{Perform kernel density estimation on SpatialExperiment for cell types of interest}
\usage{
gridDensity(
  spe,
  coi = "all",
  id = "cell_type",
  kernel = "gaussian",
  bandwidth = NULL,
  scale = 10000,
  ngrid_x = 100,
  ngrid_y = NULL,
  grid_length_x = NULL,
  grid_length_y = NULL
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs). Default to all cell types.}

\item{id}{A character. The name of the column of colData(spe) containing the cell type identifiers.
Set to cell_type by default.}

\item{kernel}{The smoothing kernel. Options are gaussian, epanechnikov, quartic or disc.}

\item{bandwidth}{The smoothing bandwidth. By default performing automatic bandwidth
selection using cross-validation using function spatstat.explore::bw.diggle.}

\item{scale}{A munmeric vector to scale the density values.}

\item{ngrid_x}{Number of grids in the x-direction. Default to 100.}

\item{ngrid_y}{Number of grids in the y-direction.}

\item{grid_length_x}{Grid length in the x-direction.}

\item{grid_length_y}{Grid length in the y-direction.}
}
\value{
A SpatialExperiment object. Grid density estimates for all cell type of interest are stored in spe@metadata$grid_density. Grid information is stored in spe@metadata$grid_info
}
\description{
Perform kernel density estimation on SpatialExperiment for cell types of interest
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

}
