% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCellCompo.R
\name{plotCellCompo}
\alias{plotCellCompo}
\title{Plot cell type composition in each density level of cell of interest.}
\usage{
plotCellCompo(
  spe,
  coi,
  id = "cell_type",
  level.name = paste0(janitor::make_clean_names(coi), "_contour"),
  by.roi = FALSE,
  all = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{id}{A character. The name of the column of colData(spe)
containing the cell type identifiers. Set to cell_type by default.}

\item{level.name}{The column name that generated by cellAssign.}

\item{by.roi}{Logical. Default is FALSE, set to TRUE to allow
plotting by ROI.}

\item{all}{Logical. Whether to include all the cell types in the plot.
Default to TRUE. If FALSE, the cell types specified in 'coi' will not
be included in the plot.}
}
\value{
A ggplot object.
}
\description{
Plot cell type composition in each density level of cell of interest.
}
\examples{

data("xenium_bc_spe")
spe <- gridDensity(spe, coi = c("Breast cancer", "Fibroblasts"))
spe <- findROI(spe, coi = c("Breast cancer", "Fibroblasts"))
spe <- getContour(spe, coi = "Breast cancer")
spe <- allocateCells(spe)
plotCellCompo(spe, coi = "Breast cancer")
plotCellCompo(spe, coi = "Breast cancer", by.roi = TRUE)

}
