% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContour.R
\name{plotContour}
\alias{plotContour}
\title{Plot contour lines.}
\usage{
plotContour(
  spe,
  coi,
  overlay = c("cell", "density"),
  sub_level = NULL,
  seed = 66,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of cell types of interest (COIs).}

\item{overlay}{Character vector. Either plot overlay on density or cell.
By default is cell.}

\item{sub_level}{Character vector. Subset on specific level.}

\item{seed}{Integer. Used for random color selection.}

\item{...}{Aesthetic mappings to pass to \code{ggplot2::aes_string()}.}
}
\value{
A ggplot object.
}
\description{
Plot contour lines.
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

coi <- "Breast cancer"

spe <- getContour(spe, coi = coi)

plotContour(spe, coi = coi, size = 0.3, alpha = 0.2)

}
