% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotContour.R
\name{plotContour}
\alias{plotContour}
\title{Plot contour lines.}
\usage{
plotContour(
  spe,
  coi,
  overlay = c("cell", "density"),
  id = "cell_type",
  sub.level = NULL,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{coi}{A character vector of length 1 of the cell type of interest (COIs).}

\item{overlay}{Character vector. Either plot overlay on density or cell.
By default is cell.}

\item{id}{A character. The name of the column of colData(spe) containing
the cell type identifiers. Set to cell_type by default.}

\item{sub.level}{Character vector. Subset on specific level.}

\item{...}{Aesthetic mappings to pass to \code{ggplot2::aes_string()}.}
}
\value{
A ggplot object.
}
\description{
Plot contour lines.
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

coi <- "Breast cancer"

spe <- getContour(spe, coi = coi)

plotContour(spe, coi = coi, size = 0.3, alpha = 0.2)

}
