% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorHeatmap.R
\name{plotCorHeatmap}
\alias{plotCorHeatmap}
\title{Plot model statistics using heatmap.}
\usage{
plotCorHeatmap(
  model.result,
  stats = c("cor.coef", "t", "p.Pos", "p.Neg"),
  roi = NULL,
  mode = c("withinROI", "acrossROI")
)
}
\arguments{
\item{model.result}{A data.frame object.}

\item{stats}{Character value. Choose either coefficient or t.
Coefficient by default.}

\item{roi}{Specify the ROI to be plotted. Leave it as default if modelling
is not done by per ROI or just want to first ROI to be plotted.}

\item{mode}{Character vector. Choose either withinROI or acrossROI.}
}
\value{
A pheatmap object.
}
\description{
Plot model statistics using heatmap.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts", "B cells", "T cells")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

spe <- corDensity(spe, cal_all = TRUE)

plotCorHeatmap(spe)

}
