% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDensCor.R
\name{plotDensCor}
\alias{plotDensCor}
\title{Plot density correlation between two cell types}
\usage{
plotDensCor(
  spe,
  celltype1 = NULL,
  celltype2 = NULL,
  by_roi = TRUE,
  ngrid = 20,
  fit = c("spline", "linear"),
  df = 3,
  ...
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{celltype1}{Cell type 1 to compare.}

\item{celltype2}{Cell type 2 to compare.}

\item{by_roi}{Logical. Plot facet by ROIs or not.}

\item{ngrid}{Integer. Threshold (minimum number of grids) used to filter small ROIs. Default to 20.}

\item{fit}{Character. Options are "spline" and "linear".}

\item{df}{Integer. Degrees of freedom of the spline fit. Default to 3 (i.e., a cubic spline fit).}

\item{...}{aesthetic mappings to pass to \code{ggplot2::aes()}.}
}
\value{
A ggplot object.
}
\description{
Plot density correlation between two cell types
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap")

plotDensCor(spe, celltype1 = "Breast cancer", celltype2 = "Fibroblasts")

}
