% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDots.R
\name{plotDots}
\alias{plotDots}
\title{Dot plot of gene expression by groups}
\usage{
plotDots(
  spe,
  feature,
  assay = "counts",
  group.by = "cell_type",
  detection.limit = 0,
  expression.limit = c(-Inf, Inf),
  scale = TRUE,
  cols = NULL,
  dot.scale = 6,
  flip.axes = FALSE
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{feature}{vector of feature names.}

\item{assay}{Name of assay to use for plotting feature.}

\item{group.by}{values to group plot by. Must be in colData of spe.}

\item{detection.limit}{threshold for minimum expression value for percentage
expression calculation (dot size)}

\item{expression.limit}{Upper and lower bound for average expression. Values
beyond this range are snapped to this limit. If only one value is provided,
it it taken as the upper bound.}

\item{scale}{Whether to scale the average expression data of each feature
using \link[base]{scale}.}

\item{cols}{Custom color palette.}

\item{dot.scale}{scale the radius of the plot. See \link[ggplot2]{scale_radius}}

\item{flip.axes}{Whether to flip the axes.}
}
\description{
Visualizing average expression and percentage of cell that expression a
certain gene(s). Similar to Seurat's DotPlot.
}
\examples{

data("xenium_bc_spe")
plotDots(spe,feature = rownames(spe)[1:5])
}
