% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGrid.R
\name{plotGrid}
\alias{plotGrid}
\title{Plot grid from metadata.}
\usage{
plotGrid(
  spe,
  group.by = NULL,
  cols = NULL,
  pol.border = FALSE,
  pol.alpha = 1,
  probs = 0
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{group.by}{values to group polygons by. Must be in colData of spe.
If NULL, will try with col if available.}

\item{cols}{Colour palette. Can be a vector of colours or a function
that accepts an integer n and return n colours.}

\item{pol.border}{Boolean. Whether to draw border for each polygon.}

\item{pol.alpha}{alpha of points between 0 and 1.}

\item{probs}{Numeric value between 0 and 1, used for filtering
uninformative grid. Only applicable for continuous values.}
}
\value{
A ggplot object.
}
\description{
Plot grid from metadata.
}
\examples{

data("xenium_bc_spe")

spe <- gridDensity(spe)

plotGrid(spe, group.by = "x_grid")

}
