% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotImage.R
\name{plotImage}
\alias{plotImage}
\title{Plot background image of spe}
\usage{
plotImage(
  spe,
  image = TRUE,
  image_id = NULL,
  sample_id = NULL,
  reverseY = NULL,
  crop = TRUE
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{image}{Logical. Whether to plot image (if present).}

\item{image_id, sample_id}{sample and image identifiers for scaling factor.
See \link[SpatialExperiment]{scaleFactors} for default behavior.}

\item{reverseY}{Logical. Whether to reverse Y coordinates. Default is TRUE
if the spe contains an image (even if not plotted) and FALSE if otherwise.}

\item{crop}{Whether to crop the plot to the spots.}
}
\value{
a ggplot object if there is a valid image, else NULL.
}
\description{
Plot background image of spe
}
