% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROI.R
\name{plotROI}
\alias{plotROI}
\title{Plot ROIs on spatial.}
\usage{
plotROI(spe, id = "cell_type", show.legend = FALSE, ...)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{id}{Character. The name of the column of colData(spe) containing
the cell type identifiers. Set to cell_type by default.}

\item{show.legend}{Logical. Show legend or not.}

\item{...}{Aesthetic mappings pass for point.}
}
\value{
A ggplot object.
}
\description{
Plot ROIs on spatial.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap", steps = 5)

plotROI(spe, pt.size = 0.3, pt.alpha = 0.2)

}
