% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotROI.R
\name{plotROI}
\alias{plotROI}
\title{Plot ROIs on spatial.}
\usage{
plotROI(spe, id = "cell_type", k = 30, ngrid = 20, show.legend = FALSE, ...)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{id}{Character. The name of the column of colData(spe) containing the cell type identifiers.
Set to cell_type by default.}

\item{k}{Integer value. The number of distinct color to be generated, default is 30.}

\item{ngrid}{Integer. The threshold (minimum number of grids) used to filter small ROIs. Default to 20.}

\item{show.legend}{Logical. Show legend or not.}

\item{...}{Aesthetic mappings to pass to \code{ggplot2::aes_string()} for point.}
}
\value{
A ggplot object.
}
\description{
Plot ROIs on spatial.
}
\examples{

data("xenium_bc_spe")

coi <- c("Breast cancer", "Fibroblasts")

spe <- gridDensity(spe, coi = coi)

spe <- findROI(spe, coi = coi, method = "walktrap", steps = 5)

plotROI(spe, ngrid = 30, size = 0.3, alpha = 0.2)

}
