% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotViolin.R
\name{plotViolin}
\alias{plotViolin}
\title{Violin plot using genes or cell data}
\usage{
plotViolin(
  spe,
  feature,
  assay = "counts",
  group.by = NULL,
  type = c("raw", "log", "cpm", "logcpm"),
  point = FALSE,
  color.by = NULL,
  ncol = NULL,
  pt.size = 0.3,
  pt.alpha = 0.3,
  pt.shape = ".",
  label.y = "Expression"
)
}
\arguments{
\item{spe}{A SpatialExperiment object.}

\item{feature}{can be a vector of gene names in rownames(spe), or column
names in colData(spe) if those columns are numeric.}

\item{assay}{Name of assay to use for plotting feature.}

\item{group.by}{values to group plot by. Must be in colData of spe and must
be either factor or character.}

\item{type}{Transformation to apply for the group/feature. Options are "raw"
, "log", "cpm", "logcpm", or a function that accepts and returns a vector of
the same length.}

\item{point}{Whether to plot points.}

\item{color.by}{values to color points by. Must be in colData of spe.}

\item{ncol}{Number of column if group.by is used.}

\item{pt.size}{Size of points.}

\item{pt.alpha}{Alpha of points between 0 and 1.}

\item{pt.shape}{Shape of points.}

\item{label.y}{Label for the y-axis.}
}
\description{
Violin plot using genes or cell data
}
\examples{

data("xenium_bc_spe")
plotViolin(spe,c("cell_area","nucleus_area"),group.by="cell_type",label.y="Area")
}
