% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readST.R
\name{readVisium}
\alias{readVisium}
\title{Read Visium output into spe}
\usage{
readVisium(
  dir,
  sample_id = "sample01",
  count = NULL,
  coord = NULL,
  image = NULL,
  scale_factors = NULL
)
}
\arguments{
\item{dir}{directory containing the Visium files}

\item{sample_id}{Name of the sample.}

\item{count}{Name of the h5 file with the count assay.}

\item{coord}{Name of the csv file with the tissue coordinates}

\item{image}{Names of the image files.}

\item{scale_factors}{Names of the scale factors file}
}
\description{
Read Visium output into spe
}
