% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readST.R
\name{readVisiumHD}
\alias{readVisiumHD}
\title{Read VisiumHD output into spe}
\usage{
readVisiumHD(dir, bin = c("016um", "008um", "002um"), ...)
}
\arguments{
\item{dir}{directory containing the VisiumHD files}

\item{bin}{Which bin size to use. Options of "016um","008um", and "002um"}

\item{...}{Parameters for readVisium}
}
\description{
Read VisiumHD output into spe
}
