% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readST.R
\name{readXenium}
\alias{readXenium}
\title{Read Xenium output into spe}
\usage{
readXenium(
  dir,
  sample_id = "sample01",
  count = file.path(dir, "cell_feature_matrix.h5"),
  coord = file.path(dir, "cells.parquet"),
  image = file.path(dir, "morphology.ome.tif"),
  image_reso = 6,
  image_layer = NULL,
  feature_type = "Gene Expression"
)
}
\arguments{
\item{dir}{directory containing the Xenium files}

\item{sample_id}{Name of the sample.}

\item{count}{Name of the h5 file with the count assay.}

\item{coord}{Name of the parquet file with the tissue coordinates}

\item{image}{Names of the ome.tif image files.}

\item{image_reso}{resolution of the image to use. From 1-8 (lower = better resolution).
See https://kb.10xgenomics.com/hc/articles/11636252598925. Default to 6}

\item{image_layer}{Which layer of the tiff image to use. Default is the
middle-most layer}

\item{feature_type}{Feature type to retain. Defaults to "Gene Expression" to
exclude control codewords. Set to NULL to keep all features.}
}
\description{
Read Xenium output into spe
}
