% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRegion.R
\name{selectRegion}
\alias{selectRegion}
\title{Select region of interest from plot}
\usage{
selectRegion(data, x_col = "x", y_col = "y")
}
\arguments{
\item{data}{A data.frame object.}

\item{xcol}{Column name of the x coordinates.}

\item{ycol}{Column name of the y coordinates.}
}
\value{
A data.frame object in the global environment.
}
\description{
Select region of interest from plot
}
\examples{

data("xenium_bc_spe")

spe_b <- spe[,colData(spe)$cell_type == "B cells"]

dat <- as.data.frame(spatialCoords(spe_b))

selectRegion(dat, x_col = "x_centroid", y_col = "y_centroid")

}
