% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectRegion.R
\name{selectRegion}
\alias{selectRegion}
\title{Select region of interest from plot}
\usage{
selectRegion(data, x.col = "x", y.col = "y")
}
\arguments{
\item{data}{A data.frame object.}

\item{x.col}{Column name of the x coordinates.}

\item{y.col}{Column name of the y coordinates.}
}
\value{
A data.frame object in the global environment.
}
\description{
Select region of interest from plot
}
\examples{

data("xenium_bc_spe")

spe_b <- spe[, SummarizedExperiment::colData(spe)$cell_type == "B cells"]

dat <- as.data.frame(SpatialExperiment::spatialCoords(spe_b))

# selectRegion(dat, x.col = "x_centroid", y.col = "y_centroid")

}
