% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scider-package.R
\docType{data}
\name{xenium_bc_spe}
\alias{xenium_bc_spe}
\alias{spe}
\title{Description of the scider example datasets}
\format{
A SpatialExperiment object
}
\usage{
data("xenium_bc_spe")
}
\description{
scider-package has 1 datasets: \itemize{
\item xenium_bc_spe Example subset of a Xenium breast cancer dataset,
}
}
\examples{
data(xenium_bc_spe)
}
\keyword{internal}
